gestion des unites mesure et correction bugs

See merge request barrydevpro14/store-backend!114

+++ src/main/java/org/store/achat/application/dto/LigneCommandeAchatUpdate.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 
 /** Command record for updating a ligne d'achat: quantité, prix, traçabilité lot. */
 public record LigneCommandeAchatUpdate(
-        int quantite,
+        BigDecimal quantite,
         BigDecimal prixAchat,
         BigDecimal prixVente,
         String numeroLot,
