+++ src/main/java/org/store/stock/application/dto/AjustementStockRequest.java
@@ -6,12 +6,13 @@ import jakarta.validation.constraints.Size;
 import org.store.stock.domain.enums.MotifAjustement;
 import org.store.stock.domain.enums.TypeAjustement;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record AjustementStockRequest(
         @NotNull UUID stockId,
         @NotNull TypeAjustement type,
-        @NotNull @Positive Integer quantite,
+        @NotNull @Positive BigDecimal quantite,
         @NotNull MotifAjustement motif,
         @Size(max = 500) String commentaire
 ) {
