+++ src/main/java/org/store/stock/application/dto/LigneEntreeStockRequest.java
@@ -6,13 +6,15 @@ import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 
 import java.math.BigDecimal;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
 public record LigneEntreeStockRequest(
         @NotNull UUID productId,
         @NotNull UUID qualityId,
-        @NotNull @Positive Integer quantite,
+        @NotNull @Positive BigDecimal quantite,
         @NotNull @DecimalMin(value = "0.0", inclusive = false) BigDecimal prixAchat,
         @NotNull @DecimalMin(value = "0.0", inclusive = false) BigDecimal prixVente,
         @Size(max = 100) String numeroLot,
