+++ src/main/java/org/store/stock/application/dto/LotConsumption.java
@@ -1,7 +1,9 @@
 package org.store.stock.application.dto;
 
+import java.math.BigDecimal;
+
 public record LotConsumption(
         SortieStockResponse sortie,
-        int restantApres
+        BigDecimal restantApres
 ) {
 }
