+++ src/main/java/org/store/stock/application/dto/LotConsumptionContext.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 
 /** Paramètres communs d'une consommation FIFO : quantite cible, prix snapshot, ligne de vente optionnelle. */
 public record LotConsumptionContext(
-        int totalAConsommer,
+        BigDecimal totalAConsommer,
         BigDecimal prixVente,
         LigneCommandeVente ligneVente
 ) {
