+++ src/main/java/org/store/stock/application/dto/MouvementJournalize.java
@@ -2,11 +2,13 @@ package org.store.stock.application.dto;
 
 import org.store.stock.domain.enums.MouvementStockType;
 
+import java.math.BigDecimal;
+
 public record MouvementJournalize(
         MouvementStockType type,
-        int quantite,
-        int stockAvant,
-        int stockApres,
+        BigDecimal quantite,
+        BigDecimal stockAvant,
+        BigDecimal stockApres,
         String referenceDocument,
         String commentaire
 ) {
