+++ src/main/java/org/store/vente/application/dto/AnnulationVenteResponse.java
@@ -5,6 +5,7 @@ import org.store.vente.domain.enums.CommandeVenteStatut;
 import org.store.vente.domain.enums.MotifAnnulationVente;
 import org.store.vente.domain.model.CommandeVente;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record AnnulationVenteResponse(
@@ -14,10 +15,10 @@ public record AnnulationVenteResponse(
         MotifAnnulationVente motif,
         String commentaire,
         String dateAnnulation,
-        int totalQuantiteReinjectee,
+        BigDecimal totalQuantiteReinjectee,
         int nombreMouvementsCrees
 ) {
-    public AnnulationVenteResponse(CommandeVente commande, int totalQuantiteReinjectee, int nombreMouvementsCrees) {
+    public AnnulationVenteResponse(CommandeVente commande, BigDecimal totalQuantiteReinjectee, int nombreMouvementsCrees) {
         this(
                 commande.getId(),
                 commande.getReference(),
